+++ CHANGELOG.md
@@ -9,3 +9,4 @@ Bob Management GUI changelog
 - Initial project structure, backend only (#9)
 - Initial project stricture, frontend (#10)
 - Dockerfile and Docker-Compose to simplify deployment (#5)
+- Logger Initialization (#14)